Split context implementations and traits into separate files
