+++ src/database/context_traits/project_context_trait.rs
@@ -1,0 +1,10 @@
+use crate::api::server::server::ProjectInfo;
+use crate::database::context_traits::EntityContextTrait;
+use crate::entities::project;
+use async_trait::async_trait;
+use sea_orm::DbErr;
+
+#[async_trait]
+pub trait ProjectContextTrait: EntityContextTrait<project::Model> {
+    async fn get_project_info_by_uid(&self, uid: i32) -> Result<Vec<ProjectInfo>, DbErr>;
+}
