+++ src/database/mod.rs
@@ -1,8 +1,2 @@
-pub mod access_context;
-pub mod database_context;
-pub mod entity_context;
-pub mod in_use_context;
-pub mod project_context;
-pub mod query_context;
-pub mod session_context;
-pub mod user_context;
+pub mod context_impls;
+pub mod context_traits;
