+++ src/main.rs
@@ -6,16 +6,9 @@ mod tests;
 use crate::api::context_collection::ContextCollection;
 use crate::api::hashing_context::HashingContext;
 use crate::api::reveaal_context::ReveaalContext;
-use crate::database::access_context::AccessContext;
-use crate::database::database_context::{PostgresDatabaseContext, SQLiteDatabaseContext};
-use crate::database::in_use_context::InUseContext;
-use crate::database::project_context::ProjectContext;
-use crate::database::query_context::QueryContext;
-use crate::database::session_context::SessionContext;
-use crate::database::user_context::UserContext;
+use crate::database::context_impls::*;
+use crate::database::context_traits::DatabaseContextTrait;
 use api::server::start_grpc_server;
-use database::database_context::DatabaseContextTrait;
-use database::entity_context::EntityContextTrait;
 use dotenv::dotenv;
 use sea_orm::{ConnectionTrait, Database, DbBackend};
 use std::env;
