+++ src/tests/api/helpers.rs
@@ -11,13 +11,8 @@ use crate::api::server::server::{
     QueryRequest, QueryResponse, SimulationStartRequest, SimulationStepRequest,
     SimulationStepResponse, UserTokenResponse,
 };
-use crate::database::access_context::AccessContextTrait;
-use crate::database::entity_context::EntityContextTrait;
-use crate::database::in_use_context::InUseContextTrait;
-use crate::database::project_context::ProjectContextTrait;
-use crate::database::query_context::QueryContextTrait;
-use crate::database::session_context::SessionContextTrait;
-use crate::database::user_context::UserContextTrait;
+use crate::database::context_traits::*;
+
 use crate::entities::{access, in_use, project, query, session, user};
 use async_trait::async_trait;
 use mockall::mock;
